Try to catch failures in "reset_consoles" like we do for activate
Resetting a console might fail as well as activating or selecting one
depending on the backend implementation. For example the IPMI backend
needs to detach from a SOL session which might have terminated already
in before which we can catch and forward to the test scope.

This is untested.

Related progress issue: https://progress.opensuse.org/issues/60437