Extract all log functions into new module "log"
The module "bmwqemu" was always a not so nicely designed catch-all
ticket. We can split out the low-level logging functions which can be
also a first step towards better structuring which can also make it
easier for us to manage dependencies within the project.

All log functions in bmwqemu are currently kept in place, forwarding to
the functions in the log module accordingly.