tools: Properly call tests from root dir
I think for a long time tests have been called from the t/ (or xt/)
directory because some tests relied on that but by now we should have
fixed all tests to be able to be called from the root folder of the
project (or any folder). With this change we call prove from the root
dir which should help us to get rid of some special workarounds as well
as allow to collect coverage information consistently from test code as
well.